Fixed the css issue with spacing around the app.  Split out the nav menu into its own component.  Updated the primary font, but it isn't getting picked up.
