+++ Moda.Web/src/moda.web.reactclient/app/layout.tsx
@@ -1,41 +1,21 @@
 'use client'
 
-import './globals.css'
+import '../styles/globals.css'
 import React, { useEffect, useState } from 'react';
-import { Inter } from 'next/font/google'
-import type { MenuProps } from 'antd';
+import { Montserrat } from 'next/font/google'
 import { Breadcrumb, Layout, Menu, ConfigProvider, theme, Space, Typography } from 'antd'
-import { LaptopOutlined, NotificationOutlined, UserOutlined } from '@ant-design/icons'
 import lightTheme from './config/theme/light-theme';
 import darkTheme from './config/theme/dark-theme';
 import { Metadata } from 'next';
 import { AuthenticatedTemplate, MsalProvider } from '@azure/msal-react';
 import { msalInstance } from './services/auth';
 import AppHeader from './components/common/app-header';
+import NavMenu from './components/common/nav-menu';
 
-const { Header, Sider, Content } = Layout;
-const { Title } = Typography;
-const inter = Inter({ subsets: ['latin'] });
+const { Sider, Content } = Layout;
 
-const items2: MenuProps['items'] = [UserOutlined, LaptopOutlined, NotificationOutlined].map(
-  (icon, index) => {
-    const key = String(index + 1);
-
-    return {
-      key: `sub${key}`,
-      icon: React.createElement(icon),
-      label: `subnav ${key}`,
-
-      children: new Array(4).fill(null).map((_, j) => {
-        const subKey = index * 4 + j + 1;
-        return {
-          key: subKey,
-          label: `option${subKey}`,
-        };
-      }),
-    };
-  },
-);
+// TODO: not working.  This should be the primary font for the app.
+const primaryFont = Montserrat({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
   title: {
@@ -52,13 +32,12 @@ export default function RootLayout({
 }) {
   theme.useToken();
   const [currentTheme, setTheme] = useState('light');
-  const [collapsed, setCollapsed] = useState(false);
 
   useEffect(() => {
     async function initialize() {
       await msalInstance.initialize();
       if (!msalInstance.getActiveAccount()) {
-        await msalInstance.loginPopup()
+        await msalInstance.loginRedirect()
           .catch((e) => { console.error(`loginRedirect failed: ${e}`) });;
       }
     }
@@ -66,23 +45,15 @@ export default function RootLayout({
   }, []);
 
   return (
-    <html lang="en">
-      <body className={inter.className}>
+    <html lang="en" className={primaryFont.className}>
+      <body>
         <ConfigProvider theme={{ token: currentTheme === 'light' ? lightTheme : darkTheme }}>
           <MsalProvider instance={msalInstance}>
             <AuthenticatedTemplate>
               <Layout className="layout" style={{ minHeight: '100vh' }}>
                 <AppHeader currentTheme={currentTheme} setTheme={setTheme} />
                 <Layout>
-                  <Sider width={200} collapsedWidth={50} collapsible collapsed={collapsed} onCollapse={(value) => setCollapsed(value)}>
-                    <Menu
-                      mode="inline"
-                      defaultSelectedKeys={['1']}
-                      defaultOpenKeys={['sub1']}
-                      style={{ height: '100%', borderRight: 0 }}
-                      items={items2}
-                    />
-                  </Sider>
+                  <NavMenu />
                   <Layout style={{ padding: '0 24px 24px' }}>
                     <Breadcrumb separator='>' style={{ margin: '16px 0' }} items={[{ title: 'Home', href: '/' }]} />
                     <Content
