+++ Moda.Web/src/moda.web.reactclient/package.json
@@ -22,9 +22,5 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "5.1.3"
-  },
-  "devDependencies": {
-    "@types/styled-components": "^5.1.26",
-    "babel-plugin-styled-components": "^2.1.3"
   }
 }
