+++ Moda.Web/src/moda.web.reactclient/styles/globals.css
@@ -1,0 +1,6 @@
+html,
+body {
+  padding: 0;
+  margin: 0;
+}
+
